+++ basta/freq_fit.py
@@ -1,8 +1,10 @@
 """
-Fitting of frequencies and frequency ratios. Surface-effect corrections.
+Fitting of frequencies and frequency-dependent properties.
 """
 import numpy as np
 import itertools
+from tqdm import tqdm
+from copy import deepcopy
 from sklearn import linear_model
 from scipy.optimize import minimize
 from scipy.interpolate import CubicSpline
@@ -10,254 +12,9 @@ from scipy.interpolate import CubicSpline
 from basta import utils_seismic as su
 
 
-def ratios(obskey, obs, ratiotype, threepoint=False):
-    """
-    Routine to compute the ratios r02, r01 and r10 from oscillation
-    frequencies, and return the desired ratio sequence, both individual
-    and combined sequences.
-
-    Developers note: We currently do not store identifying information
-    of the sequence origin in the combined sequences. We rely on them
-    being constructed/sorted identically when computed.
-
-    Parameters
-    ----------
-    obskey : array
-        Harmonic degrees, radial orders and radial orders of frequencies
-    obs : array
-        Frequencies and their error, following the structure of obs
-    ratiotype : str
-        Which ratio sequence to determine, see constants.freqtypes.rtypes
-        for possible sequences.
-    threepoint : bool
-        If True, use three point definition of r01 and r10 ratios
-        instead of default five point definition.
-
-    Returns
-    -------
-    modratio : array
-        Ratio requested from `ratiotype`
-        
-    """
-    r01, r10, r02 = True, True, True
-
-    f0 = obs[0, obskey[0, :] == 0]
-    n0 = obskey[1, obskey[0, :] == 0]
-    if (len(f0) == 0) or (len(f0) != (n0[-1] - n0[0] + 1)):
-        # Missing modes detected
-        r01, r10 = None, None
-
-    f1 = obs[0, obskey[0, :] == 1]
-    n1 = obskey[1, obskey[0, :] == 1]
-    if (len(f1) == 0) or (len(f1) != (n1[-1] - n1[0] + 1)):
-        # Missing modes detected
-        r01, r10 = None, None
-
-    f2 = obs[0, obskey[0, :] == 2]
-    n2 = obskey[1, obskey[0, :] == 2]
-    if (len(f2) == 0) or (len(f2) != (n2[-1] - n2[0] + 1)):
-        r02 = None
-
-    # Two-point frequency ratio R02
-    # -----------------------------
-    if ratiotype in ["r02", "r012", "r102"] and r02 is not None:
-        lowest_n0 = (n0[0] - 1, n1[0], n2[0])
-        l0 = lowest_n0.index(max(lowest_n0))
-
-        # Find lowest indices for l = 0, 1, and 2
-        if l0 == 0:
-            i00 = 0
-            i01 = n0[0] - n1[0] - 1
-            i02 = n0[0] - n2[0] - 1
-        elif l0 == 1:
-            i00 = n1[0] - n0[0] + 1
-            i01 = 0
-            i02 = n1[0] - n2[0]
-        elif l0 == 2:
-            i00 = n2[0] - n0[0] + 1
-            i01 = n2[0] - n1[0]
-            i02 = 0
-
-        # Number of r02s
-        nn = (n0[-1], n1[-1], n2[-1] + 1)
-        ln = nn.index(min(nn))
-        if ln == 0:
-            nr02 = n0[-1] - n0[i00] + 1
-        elif ln == 1:
-            nr02 = n1[-1] - n1[i01]
-        elif ln == 2:
-            nr02 = n2[-1] - n2[i02] + 1
-
-        # R02
-        r02 = np.zeros((nr02, 4))
-        for i in range(nr02):
-            r02[i, 0] = n0[i00 + i]
-            r02[i, 3] = f0[i00 + i]
-            r02[i, 1] = f0[i00 + i] - f2[i02 + i]
-            r02[i, 1] /= f1[i01 + i + 1] - f1[i01 + i]
-
-    # Five-point frequency ratio R01
-    # ------------------------------
-    if ratiotype in ["r01", "r012", "r010"] and r01 is not None and not threepoint:
-        # Find lowest indices for l = 0 and 1
-        if n0[0] >= n1[0]:
-            i00 = 0
-            i01 = n0[0] - n1[0]
-        else:
-            i00 = n1[0] - n0[0]
-            i01 = 0
-
-        # Number of r01s
-        if n0[-1] - 1 >= n1[-1]:
-            nr01 = n1[-1] - n1[i01]
-        else:
-            nr01 = n0[-1] - n0[i00] - 1
-
-        # R01
-        r01 = np.zeros((nr01, 4))
-        for i in range(nr01):
-            r01[i, 0] = n0[i00 + i + 1]
-            r01[i, 3] = f0[i00 + i + 1]
-            r01[i, 1] = f0[i00 + i] + 6.0 * f0[i00 + i + 1] + f0[i00 + i + 2]
-            r01[i, 1] -= 4.0 * (f1[i01 + i + 1] + f1[i01 + i])
-            r01[i, 1] /= 8.0 * (f1[i01 + i + 1] - f1[i01 + i])
-
-    # Three-point frequency ratio R01
-    # -------------------------------
-    elif ratiotype in ["r01", "r012", "r010"] and r01 is not None:
-        # Find lowest indices for l = 0 and 1
-        # i01 point to one n-value lower than i00
-        if n0[0] - 1 >= n1[0]:
-            i00 = 0
-            i01 = n0[0] - n1[0] - 1
-        else:
-            i00 = n1[0] - n0[0] + 1
-            i01 = 0
-
-        # Number of r01s
-        if n0[-1] >= n1[-1]:
-            nr01 = n1[-1] - n1[i01]
-        else:
-            nr01 = n0[-1] - n0[i00] + 1
-
-        # R01
-        r01 = np.zeros((nr01, 4))
-        for i in range(nr01):
-            r01[i, 0] = n0[i00 + i]
-            r01[i, 3] = f0[i00 + i]
-            r01[i, 1] = f0[i00 + i]
-            r01[i, 1] -= (f1[i01 + i + 1] + f1[i01 + i]) / 2.0
-            r01[i, 1] /= f1[i01 + i + 1] - f1[i01 + i]
-
-    # Five point frequency ratio R10
-    # ------------------------------
-    if ratiotype in ["r10", "r102", "r010"] and r10 is not None and not threepoint:
-        # Find lowest indices for l = 0 and 1
-        if n0[0] - 1 >= n1[0]:
-            i00 = 0
-            i01 = n0[0] - n1[0] - 1
-        else:
-            i00 = n1[0] - n0[0] + 1
-            i01 = 0
-
-        # Number of r10s
-        if n0[-1] >= n1[-1]:
-            nr10 = n1[-1] - n1[i01] - 1
-        else:
-            nr10 = n0[-1] - n0[i00]
-
-        # R10
-        r10 = np.zeros((nr10, 4))
-        for i in range(nr10):
-            r10[i, 0] = n1[i01 + i + 1]
-            r10[i, 3] = f1[i01 + i + 1]
-            r10[i, 1] = f1[i01 + i] + 6.0 * f1[i01 + i + 1] + f1[i01 + i + 2]
-            r10[i, 1] -= 4.0 * (f0[i00 + i + 1] + f0[i00 + i])
-            r10[i, 1] /= -8.0 * (f0[i00 + i + 1] - f0[i00 + i])
-
-    # Three point frequency ratio R10
-    # -------------------------------
-    elif ratiotype in ["r10", "r102", "r010"] and r10 is not None:
-        # Find lowest indices for l = 0 and 1
-        if n0[0] >= n1[0]:
-            i00 = 0
-            i01 = n0[0] - n1[0]
-        else:
-            i00 = n1[0] - n0[0]
-            i01 = 0
-
-        # Number of r10s
-        if n0[-1] >= n1[-1]:
-            nr10 = n1[-1] - n1[i01]
-        else:
-            nr10 = n0[-1] - n0[i00]
-
-        # R10
-        r10 = np.zeros((nr10, 4))
-        for i in range(nr10):
-            r10[i, 0] = n1[i01 + i]
-            r10[i, 3] = f1[i01 + i]
-            r10[i, 1] = f1[i01 + i]
-            r10[i, 1] -= (f0[i00 + i] + f0[i00 + i + 1]) / 2.0
-            r10[i, 1] /= f0[i00 + i] - f0[i00 + i + 1]
-
-    if ratiotype == "r02":
-        return r02
-
-    elif ratiotype == "r01":
-        return r01
-
-    elif ratiotype == "r10":
-        return r10
-
-    elif ratiotype == "r012":
-        # Number of ratios
-        n02 = r02.shape[0]
-        n01 = r01.shape[0]
-        n012 = n01 + n02
-
-        # R012 (R01 followed by R02)
-        r012 = np.zeros((n012, 4))
-        r012[0:n01, :] = r01[:, :]
-        r012[n01 : n01 + n02, 0] = r02[:, 0] + 0.1
-        r012[n01 : n01 + n02, 1:4] = r02[:, 1:4]
-        r012 = r012[r012[:, 0].argsort()]
-        r012[:, 0] = np.round(r012[:, 0])
-
-        return r012
-
-    elif ratiotype == "r102":
-        # Number of ratios
-        n02 = r02.shape[0]
-        n10 = r10.shape[0]
-        n102 = n10 + n02
-
-        # R102 (R10 followed by R02)
-        r102 = np.zeros((n102, 4))
-        r102[0:n10, :] = r10[:, :]
-        r102[n10 : n10 + n02, 0] = r02[:, 0] + 0.1
-        r102[n10 : n10 + n02, 1:4] = r02[:, 1:4]
-        r102 = r102[r102[:, 0].argsort()]
-        r102[:, 0] = np.round(r102[:, 0])
-
-        return r102
-
-    elif ratiotype == "r010":
-        # Number of ratios
-        n01 = r01.shape[0]
-        n10 = r10.shape[0]
-        n010 = n01 + n10
-
-        # R010 (R01 followed by R10)
-        r010 = np.zeros((n010, 4))
-        r010[0:n01, :] = r01[:, :]
-        r010[n01 : n01 + n10, 0] = r10[:, 0] + 0.1
-        r010[n01 : n01 + n10, 1:4] = r10[:, 1:4]
-        r010 = r010[r010[:, 0].argsort()]
-        r010[:, 0] = np.round(r010[:, 0])
-
-        return r010
-
+"""
+Individual frequencies
+"""
 
 def make_intervals(osc, osckey, dnu=None):
     """
@@ -868,6 +625,340 @@ def apply_BG14(modkey, mod, coeffs, scalnu):
     return corosc
 
 
+"""
+Frequency ratios
+"""
+
+def compute_ratios(obskey, obs, ratiotype, nrealizations=10000, threepoint=False):
+    """
+    Routine to compute the ratios r02, r01 and r10 from oscillation
+    frequencies, and return the desired ratio sequence, both individual
+    and combined sequences, along with their covariance matrix and its
+    inverse.
+
+    Developers note: We currently do not store identifying information
+    of the sequence origin in the combined sequences. We rely on them
+    being constructed/sorted identically when computed.
+
+    Parameters
+    ----------
+    obskey : array
+        Harmonic degrees, radial orders and radial orders of frequencies.
+    obs : array
+        Frequencies and their error, following the structure of obs.
+    ratiotype : str
+        Which ratio sequence to determine, see constants.freqtypes.rtypes
+        for possible sequences.
+    nrealizations : int
+        Number of realizations of the sampling for the computation of the
+        covariance matrix.
+    threepoint : bool
+        If True, use three point definition of r01 and r10 ratios
+        instead of default five point definition.
+
+    Returns
+    -------
+    ratio : array
+        Ratio requested from `ratiotype`.
+    ratio_cov : array
+        Covariance matrix of the requested ratio.
+    ratio_covinv : array
+        The inverse of the covariance matrix of the requested ratio.
+    """
+    ratio = compute_ratioseqs(
+            obskey,
+            obs,
+            ratiotype,
+            threepoint=threepoint
+            )
+    nr = ratio.shape[0]
+
+    # Compute and store different realizations
+    nratios = np.zeros((nrealizations, nr))
+    perturb_obs = deepcopy(obs)
+    for i in tqdm(
+            range(nrealizations),
+            desc="Sampling ratio covariances",
+            ascii=True
+            ):
+        perturb_obs[0, :] = np.random.normal(obs[0, :], obs[1, :])
+        tmpratio = compute_ratioseqs(
+            obskey, perturb_obs, ratiotype, threepoint=threepoint
+        )
+        nratios[i, :] = tmpratio[:, 1]
+
+    # Compute the covariance matrix and test the convergence
+    n = int(round(nrealizations / 2))
+    cov = np.cov(nratios[0:n, :], rowvar=False)
+    ratio_cov = np.cov(nratios, rowvar=False)
+    fnorm = np.linalg.norm(ratio_cov - cov) / (nr ** 2)
+    if fnorm > 1.0e-6:
+        print("Frobenius norm %e > 1.e-6" % (fnorm))
+        print("Warning: covariance failed to converge!")
+
+    # Compute the uncertainties on ratios using covariance matrix
+    ratio[:, 2] = np.sqrt(np.diag(ratio_cov))
+
+    # We need the inverse covariance matrix for fitting
+    ratio_covinv = np.linalg.pinv(ratio_cov, rcond=1e-8)
+    return ratio, ratio_cov, ratio_covinv
+
+
+def compute_ratioseqs(obskey, obs, ratiotype, threepoint=False):
+    """
+    Routine to compute the ratios r02, r01 and r10 from oscillation
+    frequencies, and return the desired ratio sequence, both individual
+    and combined sequences.
+
+    Developers note: We currently do not store identifying information
+    of the sequence origin in the combined sequences. We rely on them
+    being constructed/sorted identically when computed.
+
+    Parameters
+    ----------
+    obskey : array
+        Harmonic degrees, radial orders and radial orders of frequencies
+    obs : array
+        Frequencies and their error, following the structure of obs
+    ratiotype : str
+        Which ratio sequence to determine, see constants.freqtypes.rtypes
+        for possible sequences.
+    threepoint : bool
+        If True, use three point definition of r01 and r10 ratios
+        instead of default five point definition.
+
+    Returns
+    -------
+    ratio : array
+        Ratio requested from `ratiotype`
+    ratio_cov : array
+        Covariance matrix of the requested ratio
+    ratio_covinv : array
+        The inverse of the covariance matrix of the requested ratio.
+    """
+    r01, r10, r02 = True, True, True
+
+    f0 = obs[0, obskey[0, :] == 0]
+    n0 = obskey[1, obskey[0, :] == 0]
+    if (len(f0) == 0) or (len(f0) != (n0[-1] - n0[0] + 1)):
+        # Missing modes detected
+        r01, r10 = None, None
+
+    f1 = obs[0, obskey[0, :] == 1]
+    n1 = obskey[1, obskey[0, :] == 1]
+    if (len(f1) == 0) or (len(f1) != (n1[-1] - n1[0] + 1)):
+        # Missing modes detected
+        r01, r10 = None, None
+
+    f2 = obs[0, obskey[0, :] == 2]
+    n2 = obskey[1, obskey[0, :] == 2]
+    if (len(f2) == 0) or (len(f2) != (n2[-1] - n2[0] + 1)):
+        r02 = None
+
+    # Two-point frequency ratio R02
+    # -----------------------------
+    if ratiotype in ["r02", "r012", "r102"] and r02 is not None:
+        lowest_n0 = (n0[0] - 1, n1[0], n2[0])
+        l0 = lowest_n0.index(max(lowest_n0))
+
+        # Find lowest indices for l = 0, 1, and 2
+        if l0 == 0:
+            i00 = 0
+            i01 = n0[0] - n1[0] - 1
+            i02 = n0[0] - n2[0] - 1
+        elif l0 == 1:
+            i00 = n1[0] - n0[0] + 1
+            i01 = 0
+            i02 = n1[0] - n2[0]
+        elif l0 == 2:
+            i00 = n2[0] - n0[0] + 1
+            i01 = n2[0] - n1[0]
+            i02 = 0
+
+        # Number of r02s
+        nn = (n0[-1], n1[-1], n2[-1] + 1)
+        ln = nn.index(min(nn))
+        if ln == 0:
+            nr02 = n0[-1] - n0[i00] + 1
+        elif ln == 1:
+            nr02 = n1[-1] - n1[i01]
+        elif ln == 2:
+            nr02 = n2[-1] - n2[i02] + 1
+
+        # R02
+        r02 = np.zeros((nr02, 4))
+        for i in range(nr02):
+            r02[i, 0] = n0[i00 + i]
+            r02[i, 3] = f0[i00 + i]
+            r02[i, 1] = f0[i00 + i] - f2[i02 + i]
+            r02[i, 1] /= f1[i01 + i + 1] - f1[i01 + i]
+
+    # Five-point frequency ratio R01
+    # ------------------------------
+    if ratiotype in ["r01", "r012", "r010"] and r01 is not None and not threepoint:
+        # Find lowest indices for l = 0 and 1
+        if n0[0] >= n1[0]:
+            i00 = 0
+            i01 = n0[0] - n1[0]
+        else:
+            i00 = n1[0] - n0[0]
+            i01 = 0
+
+        # Number of r01s
+        if n0[-1] - 1 >= n1[-1]:
+            nr01 = n1[-1] - n1[i01]
+        else:
+            nr01 = n0[-1] - n0[i00] - 1
+
+        # R01
+        r01 = np.zeros((nr01, 4))
+        for i in range(nr01):
+            r01[i, 0] = n0[i00 + i + 1]
+            r01[i, 3] = f0[i00 + i + 1]
+            r01[i, 1] = f0[i00 + i] + 6.0 * f0[i00 + i + 1] + f0[i00 + i + 2]
+            r01[i, 1] -= 4.0 * (f1[i01 + i + 1] + f1[i01 + i])
+            r01[i, 1] /= 8.0 * (f1[i01 + i + 1] - f1[i01 + i])
+
+    # Three-point frequency ratio R01
+    # -------------------------------
+    elif ratiotype in ["r01", "r012", "r010"] and r01 is not None:
+        # Find lowest indices for l = 0 and 1
+        # i01 point to one n-value lower than i00
+        if n0[0] - 1 >= n1[0]:
+            i00 = 0
+            i01 = n0[0] - n1[0] - 1
+        else:
+            i00 = n1[0] - n0[0] + 1
+            i01 = 0
+
+        # Number of r01s
+        if n0[-1] >= n1[-1]:
+            nr01 = n1[-1] - n1[i01]
+        else:
+            nr01 = n0[-1] - n0[i00] + 1
+
+        # R01
+        r01 = np.zeros((nr01, 4))
+        for i in range(nr01):
+            r01[i, 0] = n0[i00 + i]
+            r01[i, 3] = f0[i00 + i]
+            r01[i, 1] = f0[i00 + i]
+            r01[i, 1] -= (f1[i01 + i + 1] + f1[i01 + i]) / 2.0
+            r01[i, 1] /= f1[i01 + i + 1] - f1[i01 + i]
+
+    # Five point frequency ratio R10
+    # ------------------------------
+    if ratiotype in ["r10", "r102", "r010"] and r10 is not None and not threepoint:
+        # Find lowest indices for l = 0 and 1
+        if n0[0] - 1 >= n1[0]:
+            i00 = 0
+            i01 = n0[0] - n1[0] - 1
+        else:
+            i00 = n1[0] - n0[0] + 1
+            i01 = 0
+
+        # Number of r10s
+        if n0[-1] >= n1[-1]:
+            nr10 = n1[-1] - n1[i01] - 1
+        else:
+            nr10 = n0[-1] - n0[i00]
+
+        # R10
+        r10 = np.zeros((nr10, 4))
+        for i in range(nr10):
+            r10[i, 0] = n1[i01 + i + 1]
+            r10[i, 3] = f1[i01 + i + 1]
+            r10[i, 1] = f1[i01 + i] + 6.0 * f1[i01 + i + 1] + f1[i01 + i + 2]
+            r10[i, 1] -= 4.0 * (f0[i00 + i + 1] + f0[i00 + i])
+            r10[i, 1] /= -8.0 * (f0[i00 + i + 1] - f0[i00 + i])
+
+    # Three point frequency ratio R10
+    # -------------------------------
+    elif ratiotype in ["r10", "r102", "r010"] and r10 is not None:
+        # Find lowest indices for l = 0 and 1
+        if n0[0] >= n1[0]:
+            i00 = 0
+            i01 = n0[0] - n1[0]
+        else:
+            i00 = n1[0] - n0[0]
+            i01 = 0
+
+        # Number of r10s
+        if n0[-1] >= n1[-1]:
+            nr10 = n1[-1] - n1[i01]
+        else:
+            nr10 = n0[-1] - n0[i00]
+
+        # R10
+        r10 = np.zeros((nr10, 4))
+        for i in range(nr10):
+            r10[i, 0] = n1[i01 + i]
+            r10[i, 3] = f1[i01 + i]
+            r10[i, 1] = f1[i01 + i]
+            r10[i, 1] -= (f0[i00 + i] + f0[i00 + i + 1]) / 2.0
+            r10[i, 1] /= f0[i00 + i] - f0[i00 + i + 1]
+
+    if ratiotype == "r02":
+        return r02
+
+    elif ratiotype == "r01":
+        return r01
+
+    elif ratiotype == "r10":
+        return r10
+
+    elif ratiotype == "r012":
+        # Number of ratios
+        n02 = r02.shape[0]
+        n01 = r01.shape[0]
+        n012 = n01 + n02
+
+        # R012 (R01 followed by R02)
+        r012 = np.zeros((n012, 4))
+        r012[0:n01, :] = r01[:, :]
+        r012[n01 : n01 + n02, 0] = r02[:, 0] + 0.1
+        r012[n01 : n01 + n02, 1:4] = r02[:, 1:4]
+        r012 = r012[r012[:, 0].argsort()]
+        r012[:, 0] = np.round(r012[:, 0])
+
+        return r012
+
+    elif ratiotype == "r102":
+        # Number of ratios
+        n02 = r02.shape[0]
+        n10 = r10.shape[0]
+        n102 = n10 + n02
+
+        # R102 (R10 followed by R02)
+        r102 = np.zeros((n102, 4))
+        r102[0:n10, :] = r10[:, :]
+        r102[n10 : n10 + n02, 0] = r02[:, 0] + 0.1
+        r102[n10 : n10 + n02, 1:4] = r02[:, 1:4]
+        r102 = r102[r102[:, 0].argsort()]
+        r102[:, 0] = np.round(r102[:, 0])
+
+        return r102
+
+    elif ratiotype == "r010":
+        # Number of ratios
+        n01 = r01.shape[0]
+        n10 = r10.shape[0]
+        n010 = n01 + n10
+
+        # R010 (R01 followed by R10)
+        r010 = np.zeros((n010, 4))
+        r010[0:n01, :] = r01[:, :]
+        r010[n01 : n01 + n10, 0] = r10[:, 0] + 0.1
+        r010[n01 : n01 + n10, 1:4] = r10[:, 1:4]
+        r010 = r010[r010[:, 0].argsort()]
+        r010[:, 0] = np.round(r010[:, 0])
+
+        return r010
+
+
+"""
+Epsilon difference fitting
+"""
 def compute_epsilon_diff(
     osckey,
     osc,
@@ -962,3 +1053,120 @@ def compute_epsilon_diff(
         mask = np.argsort(deps[3, :] + deps[2, :] * 0.1)
         deps = deps[:, mask]
     return deps
+
+
+"""
+Epsilon difference fitting
+"""
+def compute_epsilon_diff_and_cov(
+    osckey,
+    osc,
+    avgdnu,
+    seq="e012",
+    nsort=True,
+    extrapolation=False,
+    nrealisations=20000,
+    debug=False,
+):
+    """
+    Compute epsilon differences and covariances.
+
+    From Roxburgh 2016:
+     - Eq. 1 -> Epsilon(n,l)
+     - Eq. 4 -> EpsilonDifference(l=0,l={1,2})
+
+    Epsilon differences are independent of surface phase shift/outer
+    layers when the epsilons are evaluated at the same frequency. It
+    therefore relies on splining from epsilons at the observed frequencies
+    of the given degree and order to the frequency of the compared/subtracted
+    epsilon. See function `compute_epsilon_diff' for further clarification.
+
+    For MonteCarlo sampling of the covariances, it is replicated from the
+    covariance determination of frequency ratios in BASTA, (sec 4.1.3 of
+    Aguirre Børsen-Koch et al. 2022). A number of realisations of the
+    epsilon differences are drawn from random Gaussian distributions of the
+    individual frequencies within their uncertainty.
+
+
+    Parameters
+    ----------
+    osckey : array
+        Array containing the angular degrees and radial orders of the modes
+    osc : array
+        Array containing the modes (and inertias)
+    osccov : array
+        Covariances between individual frequencies
+    avgdnu : float
+        Average value of the large frequency separation
+    seq : str, optional
+        Similar to ratios, what sequence of epsilon differences to be computed.
+        Can be e01, e02 or e012 for a combination of the two first.
+    nsort : bool, optional
+        If True (default), the sequences are sorted by n-value of the frequencies. If
+        False, the entire 01 sequence is followed by the 02 sequence.
+    extrapolation : bool, optional
+        If False (default), modes outside the range of the l=0 modes are discarded to
+        avoid extrapolation.
+    nrealisations : int or bool, optional
+        If int: number of realisations used for MC-sampling the covariances
+        If bool: Whether to use MC (True) or analytic (False) deternubation
+        of covariances. If True, nrealisations of 20,000 is used.
+    debug : bool, optional
+        Print additional output and make plots for debugging (incl. a plot of the
+        correlation matrix)
+
+    Returns
+    -------
+    epsilon : array
+        Array containing the modes in the observed data
+    covinv : array
+        Covariances. The inverse.
+    """
+
+    # Step 0: Remove modes outside of l=0 range
+    if not extrapolation:
+        indall = osckey[0, :] > -1
+        ind0 = osckey[0, :] == 0
+        ind12 = osckey[0, :] > 0
+        mask = np.logical_and(
+            osc[0, ind12] < max(osc[0, ind0]), osc[0, ind12] > min(osc[0, ind0])
+        )
+        indall[ind12] = mask
+        if debug and any(mask):
+            print(
+                "The following modes have been skipped from epsilon differences to avoid extrapolation:"
+            )
+            for f, (l, n) in zip(osc[0, ~indall], osckey[:, ~indall].T):
+                print(" - (l,n,f) = ({0}, {1:02d}, {2:.3f})".format(l, n, f))
+
+        osc = osc[:, indall]
+        osckey = osckey[:, indall]
+
+    # Step 1: Compute epsilon differences
+    eps_diff = freq_fit.compute_epsilon_diff(
+        osckey, osc, avgdnu, seq=seq, nsorting=nsort
+    )
+
+    # Step 2a: Compute different perturbed realisations (Monte Carlo) for covariances
+    eps_reals = np.zeros((nrealisations, eps_diff.shape[1]))
+    perturb_osc = deepcopy(osc)
+    for i in tqdm(
+        range(nrealisations), desc="Sampling epsilon difference covariances", ascii=True
+    ):
+        perturb_osc[0][:] = np.random.normal(osc[0][:], osc[1][:])
+        perturb_eps = freq_fit.compute_epsilon_diff(
+            osckey, perturb_osc, avgdnu, seq=seq, nsorting=nsort
+        )
+        eps_reals[i, :] = perturb_eps[0]
+
+    # Step 2b: Derive covariance matrix from MC-realisations and test convergence
+    n = int(round(nrealisations / 2))
+    cov = np.cov(eps_reals[:n, :], rowvar=False)
+    covDeps = np.cov(eps_reals, rowvar=False)
+    fnorm = np.linalg.norm(covDeps - cov) / eps_diff.shape[1] ** 2
+    if fnorm > 1.0e-6:
+        print("Frobenius norm {0} > 1e-6".format(fnorm))
+        print("Warning: Covariance failed to converge")
+
+    return eps_diff, covDeps
+
